// frontend/src/api/generated/model/roomType.ts
export type RoomType = typeof RoomType[keyof typeof RoomType];


export const RoomType = {
  corridor: 'corridor',
  block: 'block',
  apartment: 'apartment',
} as const;